Django_bp/requirement.txt
django>=4.0.0,<4.1.0
djangorestframework
djangorestframework-simplejwt 
requests
django-cors-headers